main.py(work_alerts) created function for new posts

+++ python2day/work_alert/main.py
@@ -7,13 +7,16 @@ from selenium import webdriver
 from selenium.webdriver.support.ui import Select
 
 
-def get_list():
+def get_data():
 
+    if not  os.path.exists('data'):
+        os.mkdir('data')
+    
     try:
 
         options = webdriver.ChromeOptions()
         # options.add_argument("--no-sandbox")
-        options.add_argument("headless")
+        # options.add_argument("headless")
         options.add_argument("--disable-blink-features=AutomationControlled")
 
         driver = webdriver.Chrome(
@@ -31,50 +34,59 @@ def get_list():
         driver.find_element_by_css_selector('#comboe').send_keys('парсинг данных')
         time.sleep(1)
         driver.find_element_by_css_selector('.b-buttons .b-button').click()
-        time.sleep(1)
+
+        current_time = int(time.time())
+        data = []
         projects = driver.find_elements_by_css_selector('.b-post__grid')
-        projects_id = [project.find_element_by_css_selector('a').get_attribute('id') for project in projects]
-        return zip((projects_id, projects))
+        for project in projects:
+            project_id = project.find_element_by_css_selector('a').get_attribute('name')
+            project_title = project.find_element_by_css_selector('a').text
+            project_href = "https://www.fl.ru" + \
+                    project.find_element_by_css_selector('a').get_attribute('href')
+            project_price = project.find_element_by_css_selector('.b-post__price').text
+            project_descr = project.find_element_by_css_selector('.b-post__txt').text
+
+            data.append(
+                    {
+                        'id': project_id,
+                        'items': {
+                            'title': project_title,
+                            'href': project_href,
+                            'description': project_descr,
+                            'price': project_price,
+                            'time': f'{datetime.datetime.now().strftime("%d/%m/%Y %H:%M")}'
+                            
+                        }
+                    }
+                )
 
     finally:
         driver.quit()
+        return data
 
 
-def save_data(project):
+def update_posts(dict):
 
-    try:
-
-        project_id = project.find_element_by_css_selector('a').get_attribute('name')
-        project_title = project.find_element_by_css_selector('a').text
-        project_href = "https://www.fl.ru" + \
-                project.find_element_by_css_selector('a').get_attribute('href')
-        project_price = project.find_element_by_css_selector('.b-post__price').text
-        project_descr = project.find_element_by_css_selector('.b-post__txt').text
-
-        data = {
-                'id': project_id,
-                'items': {
-                    'title': project_title,
-                    'href': project_href,
-                    'description': project_descr,
-                    'price': project_price,
-                    'time': f'{datetime.datetime.now().strftime("%d/%m/%Y %H:%M")}'
-
-                    }
-            }
-    finally:
-
-        with open('data/posts.json', 'a', encoding='utf-8') as file:
+    if not os.path.isfile('data/posts.json'):
+        with open('data/posts.json', 'w') as file:
             json.dump(data, file, indent=4, ensure_ascii=False)
 
-def get_update():
+    else:
+        with open('data/posts.json') as file:
+            local_dict = json.load(file)
 
-    with open('data/posts.json', encoding='utf-8') as file:
-        projects = json.load(file)
+        new_dict = []
+        for item in dict:
+            if item not in local_dict:
+                new_dict.append(item)
 
-    for i in projects:
-        print(i['id'])
+        local_dict.extend(dict)
+        with open('data/posts.json', 'w') as file:
+            json.dump(local_dict, file, indent=4, ensure_ascii=False)
+    
+        print(new_dict)
 
 
 if __name__ == '__main__':
-    get_update()
+    update_posts(get_data())
+
